perf(service): defer heavy dependencies out of the boot path
`server.ts` mounts every route into one Hono app and the oRPC router imports all
13 route modules, so anything a route reaches at module scope is loaded at boot.
That put the AWS SDK, react-email, the agent runtime and the provider SDKs in the
eager module graph of a process serving nothing but content reads.

Deferred to first use:

- `packages/auth` magic-link now imports `resend` and the React template inside
  `sendMagicLink`, matching what `packages/api/email` already did. This was what
  pulled `@react-email/render`, `react-dom` and `css-tree` in through the auth
  config, which every service boots.
- `file.route` loads the S3 service on demand; `new S3Client()` ran at module scope.
- `agent-turn.step` loads the domain runtime inside the turn. Workflow steps are
  registered at boot, so its domain imports were eager by construction. The handler
  map is keyed by literal for the same reason and asserts the kind once loaded.
- `agent.service` is split into registration plus `writing-agent.service`.
- `embeddings/provider` loads `./openai.ts` inside `embed()`, which covers the
  indexing service and both index steps at once.
- `@chia/ai/utils` was a barrel mixing API-key crypto — imported by the request
  auth guards — with `createModel` and its three provider SDKs. `createModel` moves
  to `@chia/ai/utils/model`.
- `ai.route` reaches `createModel` and the gateway client through those seams.

Eager bundled code drops from 10.85 MB to 6.67 MB.

`ai` itself stays eager: `baseRequestSchema` needs `modelMessageSchema` to build
the `/generate` validator, which is evaluated when the route is defined.
`@ai-sdk/openai` and `@ai-sdk/anthropic` stay for a bundler reason rather than a
source one — rolldown co-locates `zod` in their chunks, so drizzle importing `zod`
retains them.
