+++ packages/ai/src/utils/index.ts
@@ -4,17 +4,22 @@ import {
   privateDecrypt,
 } from "node:crypto";
 
-import { createAnthropic } from "@ai-sdk/anthropic";
-import { createGoogleGenerativeAI } from "@ai-sdk/google";
-import { createOpenAI } from "@ai-sdk/openai";
-import type { LanguageModel } from "ai";
 import GithubSlugger from "github-slugger";
+import * as z from "zod";
 
-import type { BaseRequest } from "./types";
-import { Provider } from "./types";
-import { authTokenSchema } from "./types";
+/**
+ * This module is imported by the API-key guards, so it must stay free of provider SDKs.
+ * `createModel` lives in `./model` for that reason, and the auth-token schema is declared here
+ * rather than pulled from `./types` — that module reaches `ai` for `modelMessageSchema`.
+ */
 
 export const slugger = new GithubSlugger();
+
+export const authTokenSchema = z.object({
+  apiKey: z.string().min(1),
+});
+
+export type AuthToken = z.infer<typeof authTokenSchema>;
 
 export function generateKeys() {
   const { publicKey, privateKey } = generateKeyPairSync("rsa", {
@@ -52,40 +57,4 @@ export const verifyApiKey = (token: string, privateKey: string) => {
   return authTokenSchema.parse({
     apiKey: decodeApiKey(token, privateKey),
   });
-};
-
-export const createModel = (
-  options: Pick<BaseRequest, "model" | "authToken" | "proxyUrl">,
-  formater?: ((model: BaseRequest["model"]) => LanguageModel) | "ai-gateway-v3"
-): LanguageModel => {
-  if (formater) {
-    if (typeof formater === "function") {
-      return formater(options.model);
-    }
-    switch (formater) {
-      case "ai-gateway-v3":
-        return `${options.model.provider}/${options.model.id}`;
-      default:
-        throw new Error("Invalid formater");
-    }
-  }
-  switch (options.model.provider) {
-    case Provider.OpenAI:
-      return createOpenAI({
-        apiKey: options.authToken,
-        baseURL: options.proxyUrl,
-      })(options.model.id);
-    case Provider.Anthropic:
-      return createAnthropic({
-        apiKey: options.authToken,
-        baseURL: options.proxyUrl,
-      })(options.model.id);
-    case Provider.Google:
-      return createGoogleGenerativeAI({
-        apiKey: options.authToken,
-        baseURL: options.proxyUrl,
-      })(options.model.id);
-    default:
-      throw new Error("Invalid provider");
-  }
 };
