+++ packages/auth/src/base-auth.ts
@@ -4,18 +4,14 @@ import type { BetterAuthOptions } from "better-auth";
 import { magicLink } from "better-auth/plugins";
 import { admin } from "better-auth/plugins";
 import { organization } from "better-auth/plugins";
-import { Resend } from "resend";
 
 import { Role } from "@chia/db/types";
-import AuthEmailTemplate from "@chia/ui/features/AuthEmailTemplate";
 import { AUTH_EMAIL } from "@chia/utils/config";
 
 import { env } from "./env";
 import { useSecureCookies, getCookieDomain, X_CH_API_KEY } from "./utils";
 
 export const name = "auth-core";
-
-const resend = new Resend(env.RESEND_API_KEY);
 
 const getOrigin = (url?: string) => {
   if (!url) {
@@ -97,8 +93,19 @@ export const baseAuthConfig = {
 
   plugins: [
     magicLink({
+      /**
+       * `resend` and the React template are imported lazily: this config is part of
+       * every service's boot path, and a static import would pull `@react-email/render`
+       * plus the whole component library into the eager module graph for a code path
+       * that only runs when someone requests a magic link.
+       */
       sendMagicLink: async ({ email, url }) => {
-        await resend.emails.send({
+        const [{ Resend }, { default: AuthEmailTemplate }] = await Promise.all([
+          import("resend"),
+          import("@chia/ui/features/AuthEmailTemplate"),
+        ]);
+
+        await new Resend(env.RESEND_API_KEY).emails.send({
           from: AUTH_EMAIL,
           to: email,
           subject: "Sign in to Chia1104.dev",
